+++ ui/components/checkbox/RELEASENOTES.md
@@ -2,7 +2,11 @@
 
 # Checkbox Release Notes
 
-<!-- ## [Unreleased] -->
+## 2.8.0
+
+### Added
+- Added unchecked image based checkbox example [view mode checkbox - unchecked](components/checkbox/?state=view-mode-unchecked&variant=form-element)
+- Added checked image based checkbox example [view mode checkbox - checked](components/checkbox/?state=view-mode-checked&variant=form-element)
 
 ## 2.7.0
 
@@ -10,5 +14,4 @@
 - Added variant for a checkbox inside of a standard [form element component](/components/checkbox/?variant=form-element)
 
 ### Fixed
-
 - Fixed [issue #556](https://github.com/salesforce-ux/design-system/issues/556) by increasing the font size on checkbox labels to match radio button labels